elfmalloc: Eliminate dependency on Creek
This involves a refactor of the way we determine whether an object was
allocated from the Slag subsystem. A side-effect is cleaning up a lot of
the code that was strewn across slag.rs and general.rs.

One remaining problem is freeing memory when a locally-scoped allocator
is cleaned up when it is not backed by the creek.
